Fix issue with slurm-mail sending emails for each job of an array job rather than a summary when not using ARRAY_TASKS mail-type option
From the manual page for sbatch:
"Unless the ARRAY_TASKS option is specified, mail notifications on job
BEGIN, END and FAIL apply to a job array as a whole rather than
generating individual email messages for  each  task in the job array.

When the ARRAY_TASKS mail option is used in conjunction with begin,
end, or fail one email should be sent for each array task for each
condition. A job with --array=1_3 and mail-type=begin,end,array_tasks
should send 3 begin emails 3 end emails. This was working as expected.

When ARRAY_TASKS mail option is not used, there should be one email
sent when the first job of an array begins and one email when the last
array job completes. This was not working as expected.

Prior to this commit, slurm-mail was not handling array emails as
described in the sbatch manual or as it is described by the "note"
in the array began template. When not using ARRAY_TASKS, slurm-mail
was sending one email for each job begin and each job end in an array.
It should have been sending gone begin email and one end email.

The note in the begin array template states:
"Note: you will not receive e-mail notifications when other jobs in
your job array start. However, you will receive individual job
completion e-mails for each job in the array if you have requested
job completion e-mails in your job submission."
This is not true as slurm-mail was sending a begin email for each
array job that started when using --array=1-2 --mail-type=begin,end.